Update PCA pages deletes (#89)
1) Replace try/catch for cascade delete with if (records.Any())
2) Add test for dependent records to PCACommentType, PulseRouteType, TempRouteType, O2DeliveryType, BloodPressureRouteType, PainRating

+++ HITforISProject-master/IS_Proj_HIT/Pages/PCA/BloodPressureRouteType/Delete.cshtml.cs
@@ -27,6 +27,9 @@ namespace IS_Proj_HIT
 
         public async Task<IActionResult> OnGetAsync(byte? id)
         {
+            ViewData["RegularMessage"] = "Are you sure you want to delete this?";
+            ViewData["ErrorMessage"] = "";
+
             if (id == null)
             {
                 return NotFound();
@@ -52,6 +55,16 @@ namespace IS_Proj_HIT
 
             if (BloodPressureRouteType != null)
             {
+                // See if any PCA records exist with this type
+                bool usingExists = _context.Pcarecord.Any(p => p.BloodPressureRouteTypeId == BloodPressureRouteType.BloodPressureRouteTypeId);
+                if (usingExists)
+                {
+                    Console.WriteLine("PCA records exist using this record.");
+                    ViewData["RegularMessage"] = "";
+                    ViewData["ErrorMessage"] = "PCA records exist using this record. Delete not available.";
+                    return Page();
+                }
+
                 _context.BloodPressureRouteType.Remove(BloodPressureRouteType);
                 await _context.SaveChangesAsync();
             }
