+++ HITforISProject-master/IS_Proj_HIT/Pages/PCA/TempRouteType/Delete.cshtml.cs
@@ -28,6 +28,9 @@ namespace IS_Proj_HIT
 
         public async Task<IActionResult> OnGetAsync(int? id)
         {
+            ViewData["RegularMessage"] = "Are you sure you want to delete this?";
+            ViewData["ErrorMessage"] = "";
+
             if (id == null)
             {
                 return NotFound();
@@ -53,6 +56,16 @@ namespace IS_Proj_HIT
 
             if (TempRouteType != null)
             {
+                // See if any PCA records exist with this temp route type
+                bool usingExists = _context.Pcarecord.Any(p => p.TempRouteTypeId == TempRouteType.TempRouteTypeId);
+                if (usingExists)
+                {
+                    Console.WriteLine("PCA records exist using this record.");
+                    ViewData["RegularMessage"] = "";
+                    ViewData["ErrorMessage"] = "PCA records exist using this record. Delete not available.";
+                    return Page();
+                }
+
                 _context.TempRouteType.Remove(TempRouteType);
                 await _context.SaveChangesAsync();
             }
